requirements.txt
Django==1.8.2
South==1.0.2
django-bootstrap3==6.1.0
django-crispy-forms==1.4.0
django-registration-redux==1.2
psycopg2==2.6.1
